+++ ThirdPartyAdapters/adcolony/adcolony/src/main/java/com/jirbo/adcolony/AdColonyAdapter.java
@@ -5,18 +5,21 @@ import android.os.Bundle;
 import android.text.TextUtils;
 import android.util.Log;
 import android.view.View;
+import androidx.annotation.NonNull;
 import com.adcolony.sdk.AdColony;
 import com.adcolony.sdk.AdColonyAdSize;
 import com.adcolony.sdk.AdColonyAdView;
 import com.adcolony.sdk.AdColonyInterstitial;
 import com.google.ads.mediation.adcolony.AdColonyAdapterUtils;
 import com.google.ads.mediation.adcolony.AdColonyMediationAdapter;
+import com.google.android.gms.ads.AdError;
 import com.google.android.gms.ads.AdSize;
 import com.google.android.gms.ads.mediation.MediationAdRequest;
 import com.google.android.gms.ads.mediation.MediationBannerAdapter;
 import com.google.android.gms.ads.mediation.MediationBannerListener;
 import com.google.android.gms.ads.mediation.MediationInterstitialAdapter;
 import com.google.android.gms.ads.mediation.MediationInterstitialListener;
+import com.jirbo.adcolony.AdColonyManager.InitializationListener;
 import java.util.ArrayList;
 
 /**
@@ -67,37 +70,39 @@ public class AdColonyAdapter extends AdColonyMediationAdapter
   //region MediationInterstitialAdapter methods.
   @Override
   public void requestInterstitialAd(Context context,
-      MediationInterstitialListener mediationInterstitialListener,
+      final MediationInterstitialListener mediationInterstitialListener,
       Bundle serverParams,
       MediationAdRequest mediationAdRequest,
       Bundle mediationExtras) {
 
-    ArrayList<String> zoneList =
-        AdColonyManager.getInstance().parseZoneList(serverParams);
-    String requestedZone =
-        AdColonyManager.getInstance().getZoneFromRequest(zoneList, mediationExtras);
+    ArrayList<String> zoneList = AdColonyManager.getInstance().parseZoneList(serverParams);
+    final String requestedZone = AdColonyManager.getInstance()
+        .getZoneFromRequest(zoneList, mediationExtras);
     if (TextUtils.isEmpty(requestedZone)) {
-      String errorMessage = createAdapterError(ERROR_REQUEST_INVALID,
-          "Failed to request ad: zone ID is null or empty.");
+      String errorMessage = createAdapterError(ERROR_INVALID_SERVER_PARAMETERS,
+          "Missing or invalid Zone ID.");
       Log.e(TAG, errorMessage);
       mediationInterstitialListener.onAdFailedToLoad(this,
-          ERROR_REQUEST_INVALID);
+          ERROR_INVALID_SERVER_PARAMETERS);
       return;
     }
+
     adColonyInterstitialListener = new AdColonyAdListener(this, mediationInterstitialListener);
 
-    // Initialize AdColony.
-    boolean success = AdColonyManager.getInstance()
-        .configureAdColony(context, serverParams, mediationAdRequest, mediationExtras);
+    // Configures the AdColony SDK, which also initializes the SDK if it has not been yet.
+    AdColonyManager.getInstance().configureAdColony(context, serverParams, mediationAdRequest,
+        new InitializationListener() {
+          @Override
+          public void onInitializeSuccess() {
+            AdColony.requestInterstitial(requestedZone, adColonyInterstitialListener);
+          }
 
-    if (!success) {
-      String errorMessage = createAdapterError(ERROR_ADCOLONY_NOT_INITIALIZED,
-          "Failed to configure AdColony SDK.");
-      Log.w(TAG, errorMessage);
-      mediationInterstitialListener
-          .onAdFailedToLoad(this, ERROR_ADCOLONY_NOT_INITIALIZED);
-      return;
-    }
+          @Override
+          public void onInitializeFailed(@NonNull AdError error) {
+            Log.w(TAG, error.getMessage());
+            mediationInterstitialListener.onAdFailedToLoad(AdColonyAdapter.this, error);
+          }
+        });
 
     AdColony.requestInterstitial(requestedZone, adColonyInterstitialListener);
   }
@@ -122,21 +127,21 @@ public class AdColonyAdapter extends AdColonyMediationAdapter
 
   //region MediationBannerAdapter methods.
   @Override
-  public void requestBannerAd(Context context, MediationBannerListener mediationBannerListener,
+  public void requestBannerAd(Context context,
+      final MediationBannerListener mediationBannerListener,
       Bundle serverParams, AdSize adSize,
       MediationAdRequest mediationAdRequest, Bundle mediationExtras) {
 
     if (adSize == null) {
-      String errorMessage = createAdapterError(ERROR_REQUEST_INVALID,
+      String errorMessage = createAdapterError(ERROR_INVALID_SERVER_PARAMETERS,
           "Fail to request banner ad: adSize is null.");
       Log.e(TAG, errorMessage);
-      mediationBannerListener.onAdFailedToLoad(this, ERROR_REQUEST_INVALID);
+      mediationBannerListener.onAdFailedToLoad(this, ERROR_INVALID_SERVER_PARAMETERS);
       return;
     }
 
-    AdColonyAdSize adColonyAdSize =
-        AdColonyAdapterUtils.adColonyAdSizeFromAdMobAdSize(context, adSize);
-
+    final AdColonyAdSize adColonyAdSize = AdColonyAdapterUtils
+        .adColonyAdSizeFromAdMobAdSize(context, adSize);
     if (adColonyAdSize == null) {
       String errorMessage = createAdapterError(ERROR_BANNER_SIZE_MISMATCH,
           "Failed to request banner with unsupported size: " + adSize.toString());
@@ -147,37 +152,37 @@ public class AdColonyAdapter extends AdColonyMediationAdapter
 
     ArrayList<String> zoneList =
         AdColonyManager.getInstance().parseZoneList(serverParams);
-    String requestedZone =
+    final String requestedZone =
         AdColonyManager.getInstance().getZoneFromRequest(zoneList, mediationExtras);
 
     if (TextUtils.isEmpty(requestedZone)) {
-      String errorMessage = createAdapterError(ERROR_REQUEST_INVALID,
+      String errorMessage = createAdapterError(ERROR_INVALID_SERVER_PARAMETERS,
           "Failed to request ad: zone ID is null or empty");
       Log.e(TAG, errorMessage);
-      mediationBannerListener.onAdFailedToLoad(this, ERROR_REQUEST_INVALID);
+      mediationBannerListener.onAdFailedToLoad(this, ERROR_INVALID_SERVER_PARAMETERS);
       return;
     }
 
     adColonyBannerAdListener = new AdColonyBannerAdListener(this, mediationBannerListener);
 
-    // Initialize AdColony.
-    boolean success = AdColonyManager.getInstance()
-        .configureAdColony(context, serverParams, mediationAdRequest, mediationExtras);
+    // Configures the AdColony SDK, which also initializes the SDK if it has not been yet.
+    AdColonyManager.getInstance().configureAdColony(context, serverParams, mediationAdRequest,
+        new InitializationListener() {
+          @Override
+          public void onInitializeSuccess() {
+            String logMessage = String
+                .format("Requesting banner with ad size: %dx%d", adColonyAdSize.getWidth(),
+                    adColonyAdSize.getHeight());
+            Log.d(TAG, logMessage);
+            AdColony.requestAdView(requestedZone, adColonyBannerAdListener, adColonyAdSize);
+          }
 
-    if (!success) {
-      String errorMessage = createAdapterError(ERROR_ADCOLONY_NOT_INITIALIZED,
-          "Failed to configure AdColony SDK");
-      Log.w(TAG, errorMessage);
-      mediationBannerListener
-          .onAdFailedToLoad(this, ERROR_ADCOLONY_NOT_INITIALIZED);
-      return;
-    }
-
-    String logMessage = String
-        .format("Requesting banner with ad size: %dx%d", adColonyAdSize.getWidth(),
-            adColonyAdSize.getHeight());
-    Log.d(TAG, logMessage);
-    AdColony.requestAdView(requestedZone, adColonyBannerAdListener, adColonyAdSize);
+          @Override
+          public void onInitializeFailed(@NonNull AdError error) {
+            Log.w(TAG, error.getMessage());
+            mediationBannerListener.onAdFailedToLoad(AdColonyAdapter.this, error);
+          }
+        });
   }
 
   @Override